# src/main/resources/static/hackhub-openapi.yaml
openapi: 3.0.0
info:
  title: hackhub-openapi
  version: 1.0.0
paths:
...
  /api/hackathons/:hackathonId/support-requests:
    post:
      summary: Crea una nuova richiesta di supporto
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_crea_una_nuova_richiesta_di_supporto_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  titolo:
                    type: string
                  descrizione:
                    type: string
              example:
                id: ''
                titolo: ''
                descrizione: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 2df73838-d0ae-4722-ae0f-6cb5663bcf06
      requestBody:
        $ref: '#/components/requestBodies/crea_una_nuova_richiesta_di_supporto'
    get:
      summary: Restituisce tutte le richieste di supporto di un hackathon
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_restituisce_tutte_le_richieste_di_supporto_di_un_hackathon_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    titolo:
                      type: string
                    descrizione:
                      type: string
              example:
                - id: ''
                  titolo: ''
                  descrizione: ''
      parameters:
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: d7e02170-0177-4233-a400-797382d0e17a
  /api/hackathons/:hackathonId/support-requests/call?teamName=altroteam&mentoreId=b188a13d-bf42-4ea6-9ba0-5cd455fc950f&dateTime=2026-06-25T10:00:00:
    post:
      summary: Fissa una call tra mentore e team
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_fissa_una_call_tra_mentore_e_team_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: teamName
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: altroteam
        - name: mentoreId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: b188a13d-bf42-4ea6-9ba0-5cd455fc950f
        - name: dateTime
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: '2026-06-25T10:00:00'
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 2df73838-d0ae-4722-ae0f-6cb5663bcf06
  /api/hackathons/:hackathonId/support-requests/risposta?teamName=piedi&mentoreId=1c27ef8a-8b4b-46c2-aaf6-2ab932d800a0&message=come va?:
    post:
      summary: Invia una risposta testuale a una richiesta di supporto
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_invia_una_risposta_testuale_a_una_richiesta_di_supporto_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: teamName
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: piedi
        - name: mentoreId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: 1c27ef8a-8b4b-46c2-aaf6-2ab932d800a0
        - name: message
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: come va?
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: d7e02170-0177-4233-a400-797382d0e17a
  /api/hackathons/:hackathonId/support-requests/disponibilita?utenteId=bb45f4f3-48b6-4da0-b76d-c2d99e68e84c:
    put:
      summary: Registra la disponibilità di un utente
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_registra_la_disponibilit_di_un_utente_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: utenteId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: bb45f4f3-48b6-4da0-b76d-c2d99e68e84c
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 2df73838-d0ae-4722-ae0f-6cb5663bcf06
      requestBody:
        $ref: '#/components/requestBodies/registra_la_disponibilit_di_un_utente'
  /api/hackathons/:hackathonId/support-requests/calendario?utenteId=bb45f4f3-48b6-4da0-b76d-c2d99e68e84c:
    get:
      summary: Restituisce il calendario disponibilità di un utente
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_restituisce_il_calendario_disponibilit_di_un_utente_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: string
              example:
                slots: []
      parameters:
        - name: utenteId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: bb45f4f3-48b6-4da0-b76d-c2d99e68e84c
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 2df73838-d0ae-4722-ae0f-6cb5663bcf06
  /api/hackathons/:hackathonId/support-requests/disponibilita?teamName=altroteam:
    get:
      summary: Restituisce la disponibilità di un team
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_supportrequest_restituisce_la_disponibilit_di_un_team_yml
      description: ''
      tags:
        - SupportRequest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: string
              example:
                slots: []
      parameters:
        - name: teamName
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: altroteam
        - name: hackathonId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 2df73838-d0ae-4722-ae0f-6cb5663bcf06
...
servers:
  - url: http://localhost:8080
    description: Server di sviluppo
components:
  schemas:
...
    crea_una_nuova_richiesta_di_supporto:
      type: object
      properties:
        nomeTeam:
          type: string
        titolo:
          type: string
        descrizione:
          type: string
      example:
        nomeTeam: altroteam
        titolo: ciao
        descrizione: vorrei più spiegazioni
    registra_la_disponibilit_di_un_utente:
      type: object
      properties:
        intervalli:
          type: array
          items:
            type: object
            properties:
              dataInizio:
                type: string
              dataFine:
                type: string
      example:
        intervalli:
          - dataInizio: '2026-06-25T09:00:00'
            dataFine: '2026-06-25T12:00:00'
          - dataInizio: '2026-06-26T14:00:00'
            dataFine: '2026-06-26T17:00:00'
...
  requestBodies:
...
    crea_una_nuova_richiesta_di_supporto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/crea_una_nuova_richiesta_di_supporto'
      description: ''
      required: true
    registra_la_disponibilit_di_un_utente:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/registra_la_disponibilit_di_un_utente'
      description: ''
      required: true
...
  securitySchemes: {}